category 파일 생성 -> product에 있던 카테고리 파일 옮김
