+++ src/main/java/starbucks3355/starbucksServer/category/dto/request/CategoryListRequestDto.java
@@ -1,4 +1,4 @@
-package starbucks3355.starbucksServer.domainProduct.dto.requestDto;
+package starbucks3355.starbucksServer.category.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
