+++ src/main/java/starbucks3355/starbucksServer/category/entity/MiddleCategoryList.java
@@ -1,10 +1,14 @@
-package starbucks3355.starbucksServer.domainProduct.entity;
+package starbucks3355.starbucksServer.category.entity;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,12 +21,14 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class TopCategoryList {
+public class MiddleCategoryList {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
-	@Column(nullable = false, length = 30, unique = true)
+	@ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
+	@JoinColumn(name = "top_id")
+	private TopCategoryList topCategoryList;
+	@Column(length = 30, unique = true)
 	private String categoryName;
-	@Column(nullable = false, length = 255)
-	private String categoryImg;
+
 }
