+++ src/main/java/starbucks3355/starbucksServer/category/repository/CategoryListRepository.java
@@ -1,9 +1,9 @@
-package starbucks3355.starbucksServer.domainProduct.repository;
+package starbucks3355.starbucksServer.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import starbucks3355.starbucksServer.domainProduct.entity.CategoryList;
+import starbucks3355.starbucksServer.category.entity.CategoryList;
 
 @Repository
 public interface CategoryListRepository extends JpaRepository<CategoryList, Long> {
